+++ CHANGES.md
@@ -163,3 +163,6 @@ Major release of chat workflow.  Most interfaces for this changed.  See migratio
  - Fix for #196, mapping of EntityJson tags are Content objects.
  - Fix for #174, Addressable object hierarchy improved
 
+# Sep 14, 2021
+
+ - Fix for #221, Reminder Bot not starting